+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v1.01
 - fastcluster
 - diagonal mean (default) and Sinkhorn-Knopp normalisation
+- submatrix for orientation instead of diagonal
 
 ### v1.00
 -